refactor onclick
