+++ src/Core/Notifications/Components/NotificationCard.tsx
@@ -15,11 +15,8 @@ import {
 } from './NotificationCardStyles';
 import { notificationsBaseKey } from '../queries/notificationQueries';
 import { useNavigate } from 'react-router';
-import { getApps } from '../../../apps/apps';
-import { AppManifest } from '../../Client/Types';
-import { CoreContext } from '../../WorkSpace/src/WorkSpaceApi/workspaceState';
-import { deref } from '@dbeining/react-atom';
 import { useLocationKey } from '../../../packages/Filter/Hooks/useLocationKey';
+import { handleActionClick } from '../../../components/ActionCenter/handleActionClick';
 
 interface NotificationCardProps {
     notification: Notification;
@@ -40,33 +37,7 @@ export const NotificationCardNew = ({
     });
 
     const navigate = useNavigate();
-    //HACK: Doesnt scale
-    const apps = new Map<string, string>();
-    apps.set('ScopeChangeControl', 'change');
     const currentLocation = useLocationKey();
-
-    async function handleNotificationClick(appName: string, identifier: string): Promise<void> {
-        const actualName = apps.get(appName);
-        if (!actualName) throw 'App not found';
-        const app = getApps().find(({ shortName }) => shortName === actualName);
-        if (!app) throw 'Not found';
-        if (currentLocation === actualName) {
-            //mount sidesheet
-            await openSidesheet(identifier, app);
-        } else {
-            //HACK: table injected in url
-            navigate(`${app.groupe}/${app.shortName}/table#${app.shortName}/${identifier}`);
-        }
-    }
-
-    async function openSidesheet(identifier: string, app: AppManifest) {
-        const { idResolver, onSelect } = deref(CoreContext)[app.shortName];
-
-        const item = idResolver && (await idResolver(identifier));
-        if (!item) return;
-
-        onSelect && onSelect(item);
-    }
 
     const isExternalApp = notification.actionType === 'URL';
 
@@ -107,9 +78,11 @@ export const NotificationCardNew = ({
                                     )?.url,
                                     '_blank'
                                 )
-                                : handleNotificationClick(
+                                : handleActionClick(
                                     notification.sourceSystem.subSystem,
-                                    notification.sourceSystem.identifier
+                                    notification.sourceSystem.identifier,
+                                    navigate,
+                                    currentLocation
                                 );
 
                             onNavigate && onNavigate();
